BlockCache::run same-pc fast-loop: fix u64→u32 cast for exec_count. Own #116: 0de565a pushed a build-fail (grep '^error|Finished' matched the error line but pipeline rc=0; sum10/fib ran on stale-cached binary → false-green). §4 verify-what-check-measures / pipe-hides-truth family n≥8. Fix: check 'cargo build; rc=$?' directly. All regressions actually held now (sum10/fib/pieces/tier0-fuzz/rung-check on the FRESH binary).

+++ rust/sharpretro-jit/src/block_cache.rs
@@ -146,7 +146,7 @@ impl BlockCache {
                 ef(sp, spp);
                 hot += 1;
             }
-            entry.exec_count = entry.exec_count.saturating_add(hot);
+            entry.exec_count = entry.exec_count.saturating_add(hot as u32);
             self.n_execs += hot as usize;
             self.n_execs += 1;
         }
